Refactored the code

+++ lib/Services/database_service.dart
@@ -4,36 +4,32 @@ import 'package:carappsl/utilties/constants.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class DatabaseService {
-  static void updateUser(User user){
+  static void updateUser(User user) {
     usersRef.document(user.id).updateData({
       'name': user.name,
       'profileImageUrl': user.profileImageUrl,
       'bio': user.bio,
       'sellVehicle': user.sellVehicle
-
     });
   }
-  
-  
-  static Future<QuerySnapshot> searchUser (String name) {
-    Future<QuerySnapshot> users = usersRef.where('name', isGreaterThanOrEqualTo: name).getDocuments();
+
+  static Future<QuerySnapshot> searchUser(String name) {
+    Future<QuerySnapshot> users =
+        usersRef.where('name', isGreaterThanOrEqualTo: name).getDocuments();
     return users;
   }
 
-  static void createPost(Post post){
-
+  static void createPost(Post post) {
     postRef.document(post.authorId).collection('userPosts').add({
       'imageUrl': post.imageUrl,
       'caption': post.caption,
       'likeCount': post.likeCOunt,
       'authorId': post.authorId,
       'timestamp': post.timestamp,
-    }
-
-    );
+    });
   }
 
-  static void followUser ({String currentUserId, String userId}){
+  static void followUser({String currentUserId, String userId}) {
     // Add users to current users Following
     followingRef
         .document(currentUserId)
@@ -49,16 +45,17 @@ class DatabaseService {
         .setData({});
   }
 
-  static void unFollowUser ({String currentUserId, String userId}){
+  static void unFollowUser({String currentUserId, String userId}) {
     // Remove users to current users Following
     followingRef
         .document(currentUserId)
         .collection('userFollowing')
         .document(userId)
-        .get().then((doc){
-          if(doc.exists){
-            doc.reference.delete();
-          }
+        .get()
+        .then((doc) {
+      if (doc.exists) {
+        doc.reference.delete();
+      }
     });
 
     //Remove Followers
@@ -66,15 +63,16 @@ class DatabaseService {
         .document(userId)
         .collection('userFollowers')
         .document(currentUserId)
-        .get().then((doc){
-      if(doc.exists){
+        .get()
+        .then((doc) {
+      if (doc.exists) {
         doc.reference.delete();
       }
     });
   }
 
-  static Future<bool> isFollowingUser({String currentUserId, String userID}) async{
-
+  static Future<bool> isFollowingUser(
+      {String currentUserId, String userID}) async {
     DocumentSnapshot followingDoc = await followersRef
         .document(userID)
         .collection('userFollowers')
@@ -84,8 +82,7 @@ class DatabaseService {
     return followingDoc.exists;
   }
 
-  static Future<int> numFollowing(String userId) async{
-
+  static Future<int> numFollowing(String userId) async {
     QuerySnapshot followingSnapshot = await followingRef
         .document(userId)
         .collection('userFollowing')
@@ -94,8 +91,7 @@ class DatabaseService {
     return followingSnapshot.documents.length;
   }
 
-  static Future<int> numFollowers(String userId) async{
-
+  static Future<int> numFollowers(String userId) async {
     QuerySnapshot followersSnapshot = await followersRef
         .document(userId)
         .collection('userFollowers')
@@ -104,96 +100,96 @@ class DatabaseService {
     return followersSnapshot.documents.length;
   }
 
-  static Future<List<Post>> getFeedPosts(String userId) async{
+  static Future<List<Post>> getFeedPosts(String userId) async {
     QuerySnapshot feedSnapshot = await feedRef
         .document(userId)
         .collection('userFeed')
         .orderBy('timestamp', descending: true)
         .getDocuments();
 
-    List<Post> posts = feedSnapshot.documents.map((doc) => Post.fromDoc((doc))).toList();
+    List<Post> posts =
+        feedSnapshot.documents.map((doc) => Post.fromDoc((doc))).toList();
 
     return posts;
   }
 
-  static Future<List<Post>>  getUserPosts(String userId) async{
+  static Future<List<Post>> getUserPosts(String userId) async {
     QuerySnapshot userPostsSnapshot = await postRef
         .document(userId)
         .collection('userPosts')
         .orderBy('timestamp', descending: true)
         .getDocuments();
 
-    List<Post> posts = userPostsSnapshot.documents.map((doc) => Post.fromDoc((doc))).toList();
+    List<Post> posts =
+        userPostsSnapshot.documents.map((doc) => Post.fromDoc((doc))).toList();
 
     return posts;
   }
 
-  static Future<User> getUserWithId(String userId) async{
+  static Future<User> getUserWithId(String userId) async {
     DocumentSnapshot userDocSnapshot = await usersRef.document(userId).get();
-    if(userDocSnapshot.exists){
+    if (userDocSnapshot.exists) {
       return User.fromDoc(userDocSnapshot);
     }
 
     return User();
   }
 
-  static void likePost({String currentUserId, Post post}){
-
-    DocumentReference postsRef = postRef.document(post.authorId).collection('userPosts').document(post.id);
-
-    postsRef.get().then((doc) {
-
-      int likeCount = doc.data['likeCount'];
-      postsRef.updateData(({'likeCount': likeCount + 1}));
-      likesRef
-      .document(post.id)
-      .collection('postLikes')
-      .document(currentUserId)
-      .setData({});
-    });
-  }
-
-  static void unLikePost({String currentUserId, Post post}){
+  static void likePost({String currentUserId, Post post}) {
     DocumentReference postsRef = postRef
         .document(post.authorId)
         .collection('userPosts')
         .document(post.id);
 
-        postsRef.get().then((doc) {
-          int likeCOunt = doc.data['likeCount'];
-          postsRef.updateData({'likeCount' : likeCOunt - 1});
+    postsRef.get().then((doc) {
+      int likeCount = doc.data['likeCount'];
+      postsRef.updateData(({'likeCount': likeCount + 1}));
+      likesRef
+          .document(post.id)
+          .collection('postLikes')
+          .document(currentUserId)
+          .setData({});
+    });
+  }
 
-          likesRef
-              .document(post.id)
-              .collection('postLikes')
-              .document(currentUserId)
-              .get()
-          .then((doc) {if(doc.exists) {
-            doc.reference.delete();
-          }});
-        });
+  static void unLikePost({String currentUserId, Post post}) {
+    DocumentReference postsRef = postRef
+        .document(post.authorId)
+        .collection('userPosts')
+        .document(post.id);
 
+    postsRef.get().then((doc) {
+      int likeCOunt = doc.data['likeCount'];
+      postsRef.updateData({'likeCount': likeCOunt - 1});
 
+      likesRef
+          .document(post.id)
+          .collection('postLikes')
+          .document(currentUserId)
+          .get()
+          .then((doc) {
+        if (doc.exists) {
+          doc.reference.delete();
+        }
+      });
+    });
   }
 
   static Future<bool> didLikePost({String currentUserId, Post post}) async {
     DocumentSnapshot userDoc = await likesRef
-    .document(post.id)
+        .document(post.id)
         .collection('postLikes')
         .document(currentUserId)
         .get();
     return userDoc.exists;
-
   }
 
-  static void commentOnPost({String currentUserId, String postId, String comment}){
+  static void commentOnPost(
+      {String currentUserId, String postId, String comment}) {
     commentRef.document(postId).collection('postComments').add({
       'content': comment,
       'authorId': currentUserId,
       'timestamp': Timestamp.fromDate(DateTime.now()),
     });
   }
-
-
-
-  }
+}
